+++ src/main/java/org/bitbucket/evl/Method3D.java
@@ -38,7 +38,7 @@ public class Method3D<ReturnType, DataType> extends InvokableMethod3D<ReturnType
 		return this;
 	}
 	
-	public Method3D<ReturnType, DataType> boundedCache(long capacity) {
+	public Method3D<ReturnType, DataType> boundedCache(int capacity) {
 		this.cache = CacheFactory.<ClassTuple, DispatchableMethodD<DataType>>createBoundedCache(capacity);
 		return this;
 	}
