+++ src/main/java/org/bitbucket/evl/Method4D.java
@@ -38,7 +38,7 @@ public class Method4D<ReturnType, DataType> extends InvokableMethod4D<ReturnType
 		return this;
 	}
 	
-	public Method4D<ReturnType, DataType> boundedCache(long capacity) {
+	public Method4D<ReturnType, DataType> boundedCache(int capacity) {
 		this.cache = CacheFactory.<ClassTuple, DispatchableMethodD<DataType>>createBoundedCache(capacity);
 		return this;
 	}
